requirements.txt
environs
python-binance
fastapi
uvicorn
